Added problem Nguyen 1,2,# benchmark problem and Wine quality symbolic regression, such as repaired mutation, elitism, reproduction, XML parser, some development print-outs (you will need to comment them manually out, mainly in the Task.cpp file), some datasets

+++ src/problems/Problem.cpp
@@ -68,7 +68,7 @@ Solution Problem::generateRandomSolution(size_t maxDepth, size_t maxNodes) {
     return solution;
 }
 
-double Problem::evaluate(Solution *solution) {
+float Problem::evaluate(Solution *solution) {
     // Default implementation
     throw std::runtime_error("Problem::evaluate() not implemented.");
 }
